+++ jest.config.js
@@ -1,3 +1,24 @@
 module.exports = {
-  preset: '@vue/cli-plugin-unit-jest'
+  preset: '@vue/cli-plugin-unit-jest',
+  "collectCoverage": true,
+  "collectCoverageFrom": ["**/*.{js,vue}", "!**/node_modules/**"],
+  "coverageReporters": ["html", "text-summary"],
+  "moduleFileExtensions": [
+    "js",
+    "json",
+    "vue"
+  ],
+  "transform": {
+    ".*\\.(vue)$": "vue-jest",
+    "^.+\\.js$": "<rootDir>/node_modules/babel-jest"
+  },
+  "snapshotSerializers": ["jest-serializer-vue"],
+  "setupFiles": ["jest-localstorage-mock","<rootDir>/__mocks__/client.js"],
+  "moduleNameMapper": {
+    "\\.(css|less)$": "<rootDir>/__mocks__/styleMock.js",
+    "\\.(gif|ttf|eot|svg)$": "<rootDir>/__mocks__/fileMock.js",
+  },
+  "resolver":null,
+  "verbose":true,
+  "testEnvironment": "jest-environment-jsdom-global"
 }