+++ CHANGELOG.md
@@ -3,9 +3,17 @@ All notable changes to this project will be documented in this file. Changes sho
 
 ## [Unreleased]
 
-## [0.0.3] - 2020-04-06
+## [0.0.5] - 2020-05-01
 ### Added
- -  [Add undo/redo buttons](fb2a4dede443afdc801ecf96c5cc2f7b01799cd5)
+ - [Incorporate minimum testing](a85fb02c1d4c6b147433b9074cdf2dc1dca62258)
+
+## [0.0.4] - 2020-04-17
+### Added
+ - [Incorporate logging into output object](bb7dd43a0b0b13118af78049509b1d702d205a32)
+
+## [0.0.3] - 2020-04-03
+### Added
+ - [Addition of undo redo](fc60bbfb1076619ae9fe081a721c5cf2b940547f)
 
 ## [0.0.2] - 2020-03-13
 ### Added